fix: rework pdk exec script
